cifsd: move connection related functions to connection.c
